requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.5.1
pytz==2020.4
Pillow==8.0.1
alembic==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.0b1
texttable==1.6.3
cffi==1.14.4
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
asyncpg==0.21.0
greenlet==1.0a1
importlib-metadata==3.3.0
zipp==3.4.0
dblpy==0.4.0
ratelimiter==1.2.0.post0